+++ src/components/ui/profileCard/profileCard.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import ProfileCardImage from './profileCardImage';
-import ProfileCardName from './profileCardName';
-import ProfileCardComment from './profileCardComment';
+import ProfileCardImage from '@/components/ui/profileCard/profileCardImage'
+import ProfileCardName from '@/components/ui/profileCard/profileCardName';
+import ProfileCardComment from '@/components/ui/profileCard/profileCardComment';
 
 const ProfileCard: React.FC = () => {
   // 배경색 목록
@@ -34,7 +34,7 @@ const ProfileCard: React.FC = () => {
   }, [previousBackground]);
 
   // 잠금 상태
-  const [isLock, setLock] = useState<boolean>(true);
+  const [isLock, ] = useState<boolean>(true);
 
   const profileCardStyle = `flex ${isLock ? 'h-[360px] my-[calc((100vh-200px-360px)/2)]' : 'h-auto my-[0px]'} mx-[4%] rounded-[16px] ${currentBackground} w-[calc(100%-8%)]`;
   // transform -translate-y-1/2
